requirements.txt
pyserial>=3.5
rich>=13.7.0
pyyaml>=6.0.2
PySide6>=6.7.0
